requirements.txt
torch>=2.0.0
torch-geometric>=2.5.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
faker>=20.0.0
tqdm>=4.65.0
